refactor: combine callouts and seperate styles
