+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "cSpell.words": [
         "Bezier",
+        "Callout",
         "ekeys",
         "frontmatter",
         "iframe",
