+++ src/components/callout/styles.js
@@ -1,10 +1,7 @@
-import React from 'react';
 import styled from 'styled-components';
+import { colors, padding } from '../../theme';
 
-import { BookOpen } from 'react-feather';
-import { colors, padding } from '../theme';
-
-const StyledCallout = styled.div`
+export const CalloutWrapper = styled.div`
   position: relative;
   p {
     color: ${colors.white};
@@ -17,41 +14,23 @@ const StyledCallout = styled.div`
   border-radius: ${padding.small};
   padding: ${padding.large};
 `;
-
-const CalloutContent = styled.div`
+export const CalloutContent = styled.div`
   margin: 0;
   p {
     margin: 0;
   }
 `;
-
-const CalloutIcon = styled.div`
+export const CalloutIcon = styled.div`
   position: absolute;
   left: -${padding.medium};
   box-shadow: 0 0 0 ${padding.xsmall} ${colors.navy};
   background: ${colors.navy};
 `;
 
-const CalloutTitle = styled.div`
+export const CalloutTitle = styled.div`
   position: absolute;
   top: -${padding.medium};
   box-shadow: 0 0 0 ${padding.xsmall} ${colors.navy};
   background: ${colors.navy};
   color: ${props => props.color};
 `;
-
-const RelatedContent = ({ children }) => {
-  return (
-    <StyledCallout color={colors.black}>
-      <CalloutTitle color={colors.green}>
-        <h2>Continue reading</h2>
-      </CalloutTitle>
-      <CalloutIcon>
-        <BookOpen size={padding.large} color={colors.green} />
-      </CalloutIcon>
-      <CalloutContent>{children}</CalloutContent>
-    </StyledCallout>
-  );
-};
-
-export default RelatedContent;
